ANCHOR-81,102,114: SEP-24 transaction data model and pass stellar-anchor-tests (#677)
* Refactored SEP-24 data model
* Fix to pass the integration test
* Pass the stellar anchor tests
* clean up and revert default config values
* Add SEP-24 to stellar anchor test as part of GH workflow.
* Clean up Sep24Helper
* Add check_certificate field to CallbackApiConfig

+++ platform/src/test/kotlin/org/stellar/anchor/platform/config/Sep24ConfigTest.kt
@@ -40,8 +40,8 @@ class Sep24ConfigTest {
 
   @ParameterizedTest
   @ValueSource(ints = [-1, Integer.MIN_VALUE, 0])
-  fun `test bad interactive jwt expiration`(expiration: Integer) {
-    config.setInteractiveJwtExpiration(expiration.toInt())
+  fun `test bad interactive jwt expiration`(expiration: Int) {
+    config.setInteractiveJwtExpiration(expiration)
     config.validate(config, errors)
     assertTrue(errors.hasErrors())
     assertErrorCode(errors, "sep24-interactive-url-invalid")
